Don't show series with a missing team
When one team has advanced to the next round but their opponent isn't known yet, we were getting TBD info from the server. It's not super helpful to see that as we can already see that a team has 4 wins from a previous round, so we know they've advanced. This will only show the next round's info when both teams are known.